src/check/requirements.txt
certifi==2018.8.13
chardet==3.0.4
idna==2.7
influxdb==5.2.0
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.5
requests==2.19.1
six==1.11.0
speedtest-cli==2.0.2
urllib3==1.23
